collector/requirements.txt
deta
ulid-py
fastapi
pydantic
